+++ src/dbt_score/__init__.py
@@ -1,6 +1,15 @@
 """Init dbt_score package."""
 
+from dbt_score.model_filter import ModelFilter, model_filter
 from dbt_score.models import Model
 from dbt_score.rule import Rule, RuleViolation, Severity, rule
 
-__all__ = ["Model", "Rule", "RuleViolation", "Severity", "rule"]
+__all__ = [
+    "Model",
+    "ModelFilter",
+    "Rule",
+    "RuleViolation",
+    "Severity",
+    "model_filter",
+    "rule",
+]