+++ src/dbt_score/rule.py
@@ -4,8 +4,9 @@ import inspect
 import typing
 from dataclasses import dataclass, field
 from enum import Enum
-from typing import Any, Callable, Type, TypeAlias, overload
+from typing import Any, Callable, Iterable, Type, TypeAlias, overload
 
+from dbt_score.model_filter import ModelFilter
 from dbt_score.models import Model
 
 
@@ -24,6 +25,7 @@ class RuleConfig:
 
     severity: Severity | None = None
     config: dict[str, Any] = field(default_factory=dict)
+    model_filter_names: list[str] = field(default_factory=list)
 
     @staticmethod
     def from_dict(rule_config: dict[str, Any]) -> "RuleConfig":
@@ -34,8 +36,15 @@ class RuleConfig:
             if "severity" in rule_config
             else None
         )
+        filter_names = (
+            config.pop("model_filter_names", None)
+            if "model_filter_names" in rule_config
+            else []
+        )
 
-        return RuleConfig(severity=severity, config=config)
+        return RuleConfig(
+            severity=severity, config=config, model_filter_names=filter_names
+        )
 
 
 @dataclass
@@ -53,6 +62,8 @@ class Rule:
 
     description: str
     severity: Severity = Severity.MEDIUM
+    model_filter_names: list[str]
+    model_filters: frozenset[ModelFilter] = frozenset()
     default_config: typing.ClassVar[dict[str, Any]] = {}
 
     def __init__(self, rule_config: RuleConfig | None = None) -> None:
@@ -83,6 +94,7 @@ class Rule:
         self.set_severity(
             rule_config.severity
         ) if rule_config.severity else rule_config.severity
+        self.model_filter_names = rule_config.model_filter_names
         self.config = config
 
     def evaluate(self, model: Model) -> RuleViolation | None:
@@ -90,9 +102,21 @@ class Rule:
         raise NotImplementedError("Subclass must implement method `evaluate`.")
 
     @classmethod
+    def should_evaluate(cls, model: Model) -> bool:
+        """Checks if all filters in the rule allow evaluation."""
+        if cls.model_filters:
+            return all(f.evaluate(model) for f in cls.model_filters)
+        return True
+
+    @classmethod
     def set_severity(cls, severity: Severity) -> None:
         """Set the severity of the rule."""
         cls.severity = severity
+
+    @classmethod
+    def set_filters(cls, model_filters: Iterable[ModelFilter]) -> None:
+        """Set the filters of the rule."""
+        cls.model_filters = frozenset(model_filters)
 
     @classmethod
     def source(cls) -> str:
@@ -118,6 +142,7 @@ def rule(
     *,
     description: str | None = None,
     severity: Severity = Severity.MEDIUM,
+    model_filters: set[ModelFilter] | None = None,
 ) -> Callable[[RuleEvaluationType], Type[Rule]]:
     ...
 
@@ -127,6 +152,7 @@ def rule(
     *,
     description: str | None = None,
     severity: Severity = Severity.MEDIUM,
+    model_filters: set[ModelFilter] | None = None,
 ) -> Type[Rule] | Callable[[RuleEvaluationType], Type[Rule]]:
     """Rule decorator.
 
@@ -140,6 +166,7 @@ def rule(
         __func: The rule evaluation function being decorated.
         description: The description of the rule.
         severity: The severity of the rule.
+        model_filters: Set of ModelFilter that filters the rule.
     """
 
     def decorator_rule(
@@ -172,6 +199,7 @@ def rule(
             {
                 "description": rule_description,
                 "severity": severity,
+                "model_filters": model_filters or frozenset(),
                 "default_config": default_config,
                 "evaluate": wrapped_func,
                 # Save provided evaluate function