+++ tests/test_evaluation.py
@@ -179,3 +179,59 @@ def test_evaluation_rule_with_config(
     )
     assert evaluation.results[model1][rule_with_config] is not None
     assert evaluation.results[model2][rule_with_config] is None
+
+
+def test_evaluation_with_filter(manifest_path, default_config, rule_with_filter):
+    """Test rule with filter."""
+    manifest_loader = ManifestLoader(manifest_path)
+    mock_formatter = Mock()
+    mock_scorer = Mock()
+
+    rule_registry = RuleRegistry(default_config)
+    rule_registry._add_rule(rule_with_filter)
+
+    # Ensure we get a valid Score object from the Mock
+    mock_scorer.score_model.return_value = Score(10, "🥇")
+
+    evaluation = Evaluation(
+        rule_registry=rule_registry,
+        manifest_loader=manifest_loader,
+        formatter=mock_formatter,
+        scorer=mock_scorer,
+    )
+    evaluation.evaluate()
+
+    model1 = manifest_loader.models[0]
+    model2 = manifest_loader.models[1]
+
+    assert rule_with_filter not in evaluation.results[model1]
+    assert isinstance(evaluation.results[model2][rule_with_filter], RuleViolation)
+
+
+def test_evaluation_with_class_filter(
+    manifest_path, default_config, class_rule_with_filter
+):
+    """Test rule with filters and filtered rules defined by classes."""
+    manifest_loader = ManifestLoader(manifest_path)
+    mock_formatter = Mock()
+    mock_scorer = Mock()
+
+    rule_registry = RuleRegistry(default_config)
+    rule_registry._add_rule(class_rule_with_filter)
+
+    # Ensure we get a valid Score object from the Mock
+    mock_scorer.score_model.return_value = Score(10, "🥇")
+
+    evaluation = Evaluation(
+        rule_registry=rule_registry,
+        manifest_loader=manifest_loader,
+        formatter=mock_formatter,
+        scorer=mock_scorer,
+    )
+    evaluation.evaluate()
+
+    model1 = manifest_loader.models[0]
+    model2 = manifest_loader.models[1]
+
+    assert class_rule_with_filter not in evaluation.results[model1]
+    assert isinstance(evaluation.results[model2][class_rule_with_filter], RuleViolation)